extract legacy proposal review mutation hook, drop test hook mocks
