+++ src/app/[orgSlug]/study/[studyId]/review/proposal-review-view.test.tsx
@@ -10,45 +10,14 @@ import {
 } from '@/tests/unit.helpers'
 import { memoryRouter } from 'next-router-mock'
 import { useParams } from 'next/navigation'
-import { beforeEach, describe, expect, it, vi } from 'vitest'
+import { beforeEach, describe, expect, it } from 'vitest'
 import { ProposalReviewView } from './proposal-review-view'
-
-const proposalReviewMocks = vi.hoisted(() => ({
-    feedback: {
-        value: '',
-        wordCount: 0,
-        isValid: false,
-    },
-    submitReview: vi.fn(),
-}))
-
-vi.mock('@/hooks/use-review-feedback', () => ({
-    useReviewFeedback: () => ({
-        value: proposalReviewMocks.feedback.value,
-        onChange: vi.fn(),
-        wordCount: proposalReviewMocks.feedback.wordCount,
-        minWords: 50,
-        maxWords: 500,
-        isValid: proposalReviewMocks.feedback.isValid,
-    }),
-}))
-
-vi.mock('@/hooks/use-proposal-review-mutation', () => ({
-    useProposalReviewMutation: () => ({
-        submitReview: proposalReviewMocks.submitReview,
-        isPending: false,
-    }),
-}))
 
 describe('ProposalReviewView', () => {
     let study: SelectedStudy
 
     beforeEach(async () => {
         memoryRouter.setCurrentUrl('/')
-        proposalReviewMocks.feedback.value = ''
-        proposalReviewMocks.feedback.wordCount = 0
-        proposalReviewMocks.feedback.isValid = false
-        proposalReviewMocks.submitReview.mockReset()
         const { org, user } = await mockSessionWithTestData({ orgSlug: 'test-org', orgType: 'enclave' })
         const { study: dbStudy } = await insertTestStudyJobData({
             org,
@@ -111,63 +80,13 @@ describe('ProposalReviewView', () => {
 
             await user.click(screen.getByRole('radio', { name: /Needs clarification/ }))
 
-            // Feedback editor is still skeleton until OTTER-491 lands, so submit stays gated on invalid feedback
             expect(screen.getByRole('button', { name: 'Submit review' })).toBeDisabled()
         })
     })
 
-    describe('submit wiring', () => {
-        beforeEach(() => {
-            proposalReviewMocks.feedback.value = 'valid feedback '.repeat(60).trim()
-            proposalReviewMocks.feedback.wordCount = 120
-            proposalReviewMocks.feedback.isValid = true
-        })
-
-        it('submits approve through the unified mutation after confirmation', async () => {
-            const user = userEvent.setup()
-            renderWithProviders(<ProposalReviewView orgSlug="test-org" study={study} />)
-
-            await user.click(screen.getByRole('radio', { name: /Approve/ }))
-            await user.click(screen.getByRole('button', { name: 'Submit review' }))
-            await user.click(screen.getByRole('button', { name: 'Yes, submit review' }))
-
-            expect(proposalReviewMocks.submitReview).toHaveBeenCalledWith({
-                decision: 'approve',
-                feedback: proposalReviewMocks.feedback.value,
-            })
-        })
-
-        it('submits needs clarification through the unified mutation after confirmation', async () => {
-            const user = userEvent.setup()
-            renderWithProviders(<ProposalReviewView orgSlug="test-org" study={study} />)
-
-            await user.click(screen.getByRole('radio', { name: /Needs clarification/ }))
-            await user.click(screen.getByRole('button', { name: 'Submit review' }))
-            await user.click(screen.getByRole('button', { name: 'Yes, submit review' }))
-
-            expect(proposalReviewMocks.submitReview).toHaveBeenCalledWith({
-                decision: 'needs-clarification',
-                feedback: proposalReviewMocks.feedback.value,
-            })
-        })
-
-        it('submits reject through the destructive modal', async () => {
-            const user = userEvent.setup()
-            renderWithProviders(<ProposalReviewView orgSlug="test-org" study={study} />)
-
-            await user.click(screen.getByRole('radio', { name: /Reject/ }))
-            await user.click(screen.getByRole('button', { name: 'Submit review' }))
-
-            expect(screen.getByText('Reject initial request?')).toBeInTheDocument()
-
-            await user.click(screen.getByRole('button', { name: 'Reject initial request' }))
-
-            expect(proposalReviewMocks.submitReview).toHaveBeenCalledWith({
-                decision: 'reject',
-                feedback: proposalReviewMocks.feedback.value,
-            })
-        })
-    })
+    // Submit-wiring coverage (approve/needs-clarification/reject → mutation call with right decision + feedback)
+    // lives in server/actions/study.actions.test.ts. The UI path will be covered here once OTTER-491
+    // replaces the feedback <Skeleton> with a real editor that tests can type into.
 
     describe('already-decided guard', () => {
         it('hides decision section and action bar when study is APPROVED', () => {
